dashboard/backend/requirements-base.txt
fastapi
paho-mqtt
asyncpg
supabase
python-dotenv
geojson
pydantic-settings
PyJWT
uvicorn
pytest
requests